UPDATE: Settings Done

+++ Models/Path/Point.py
@@ -4,8 +4,8 @@ class Point:
 
     Attributes
     ----------
-    x : int
-    y : int
+    x : float
+    y : float
     is_good : bool (Used in OBSTACLE Game Mode)
         is_good==True are the points player needs to touch. is_good==False are the points player needs to avoid.
     order: int (Used in SEQUENCE Game Mode)
@@ -13,7 +13,7 @@ class Point:
     alphabet: str (Used in ALPHABET Game Mode)
         the alphabet this point represents.
     """
-    def __init__(self, x: int, y: int, is_good:bool=True, order:int=0, alphabet:str=''):
+    def __init__(self, x: float, y: float, is_good:bool=True, order:int=0, alphabet:str=''):
         self.x = x
         self.y = y
         self.is_good = is_good
